Implement view for edit a contact
Implement the view for create, update or delete a contact
Fix bug in the view for edit an user which was showing the 'delete' button when creating an user
[Time: 1h]

+++ Project/src/java/io/github/jfmdev/jsfblackbook/beans/UserFormController.java
@@ -139,8 +139,10 @@ public class UserFormController {
      */
     public Boolean delete() {
         try {    
-            UsersDAO.delete(user.getId());
-            return true;
+            if(user.getId() != null) {
+                UsersDAO.delete(user.getId());
+                return true;
+            }
         }catch(SQLException e) {
             FacesMessage facesMsg = new FacesMessage(FacesMessage.SEVERITY_ERROR, e.getMessage(), null);
             FacesContext.getCurrentInstance().addMessage(null, facesMsg);
